+++ tests/FacebookClientTest.php
@@ -34,11 +34,12 @@ use Facebook\FileUpload\FacebookVideo;
 use Facebook\HttpClients\FacebookCurlHttpClient;
 use Facebook\HttpClients\FacebookGuzzleHttpClient;
 use Facebook\HttpClients\FacebookStreamHttpClient;
-use Facebook\Tests\Fixtures\MyFooBatchClientHandler;
-use Facebook\Tests\Fixtures\MyFooClientHandler;
+use Facebook\Tests\Fixtures\MyFooBatchHttpClient;
+use Facebook\Tests\Fixtures\MyFooHttpClient;
 use Facebook\FacebookResponse;
 use Facebook\FacebookBatchResponse;
 use Facebook\GraphNodes\GraphNode;
+use Http\Client\HttpClient;
 use PHPUnit\Framework\TestCase;
 
 class FacebookClientTest extends TestCase
@@ -66,28 +67,24 @@ class FacebookClientTest extends TestCase
     protected function setUp()
     {
         $this->fbApp = new FacebookApp('id', 'shhhh!');
-        $this->fbClient = new FacebookClient(new MyFooClientHandler());
+        $this->fbClient = new FacebookClient(new MyFooHttpClient());
     }
 
     public function testACustomHttpClientCanBeInjected()
     {
-        $handler = new MyFooClientHandler();
+        $handler = new MyFooHttpClient();
         $client = new FacebookClient($handler);
-        $httpHandler = $client->getHttpClientHandler();
+        $httpClient = $client->getHttpClient();
 
-        $this->assertInstanceOf(MyFooClientHandler::class, $httpHandler);
+        $this->assertInstanceOf(MyFooHttpClient::class, $httpClient);
     }
 
     public function testTheHttpClientWillFallbackToDefault()
     {
         $client = new FacebookClient();
-        $httpHandler = $client->getHttpClientHandler();
+        $httpClient = $client->getHttpClient();
 
-        if (function_exists('curl_init')) {
-            $this->assertInstanceOf(FacebookCurlHttpClient::class, $httpHandler);
-        } else {
-            $this->assertInstanceOf(FacebookStreamHttpClient::class, $httpHandler);
-        }
+        $this->assertInstanceOf(HttpClient::class, $httpClient);
     }
 
     public function testBetaModeCanBeDisabledOrEnabledViaConstructor()
@@ -143,7 +140,7 @@ class FacebookClientTest extends TestCase
         ];
         $fbBatchRequest = new FacebookBatchRequest($this->fbApp, $fbRequests);
 
-        $fbBatchClient = new FacebookClient(new MyFooBatchClientHandler());
+        $fbBatchClient = new FacebookClient(new MyFooBatchHttpClient());
         $response = $fbBatchClient->sendBatchRequest($fbBatchRequest);
 
         $this->assertInstanceOf(FacebookBatchResponse::class, $response);
@@ -281,18 +278,6 @@ class FacebookClientTest extends TestCase
             FacebookTestCredentials::$appSecret
         );
 
-        // Use default client
-        $client = null;
-
-        // Uncomment to enable curl implementation.
-        //$client = new FacebookCurlHttpClient();
-
-        // Uncomment to enable stream wrapper implementation.
-        //$client = new FacebookStreamHttpClient();
-
-        // Uncomment to enable Guzzle implementation.
-        //$client = new FacebookGuzzleHttpClient();
-
-        static::$testFacebookClient = new FacebookClient($client);
+        static::$testFacebookClient = new FacebookClient();
     }
 }